llm-monitoring/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# LLM API clients
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
perplexity-client==0.1.0
aiohttp==3.9.1
httpx==0.25.2

# Task queue and caching
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# NLP and text processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
transformers==4.36.2
sentence-transformers==2.2.2

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
motor==3.3.2

# Monitoring and observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Utils
python-multipart==0.0.6
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
ratelimit==2.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Webhook handling
pywebhook==0.1.0
cryptography==41.0.7

# Data processing
pandas==2.1.4
numpy==1.26.2